+++ src/generator/from_proto.rs
@@ -381,7 +381,7 @@ mod test {
 
     use prost_reflect::prost_types::{FileDescriptorProto, FileDescriptorSet};
 
-    use crate::config_generator::from_proto::{from_proto, Context, DescriptorType};
+    use crate::generator::from_proto::{from_proto, Context, DescriptorType};
 
     fn get_proto_file_descriptor(name: &str) -> anyhow::Result<FileDescriptorProto> {
         let mut proto_path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
