+++ src/generator/proto/greetings_a.proto
