+++ src/generator/proto/greetings_b.proto
