+++ src/generator/proto/news.proto
