+++ src/generator/proto/news_no_service.proto
